requirements.txt
requests
google-generativeai
chromadb
